# cuda: 1.2x faster dequantization kernel