@YellowRoseCx 2023-08-26
Have you tested to see if this breaks CUDA compatibility for AMD cards using the recently merged ROCm pull request?