@KerfuffleV2 2023-08-26
> Have you tested to see if this breaks CUDA compatibility for AMD cards using the recently merged ROCm pull request?

I tried it out:

```plaintext
ggml_init_cublas: found 1 ROCm devices:
  Device 0: AMD Radeon RX 6600, compute capability 10.3
```

Seems fine, identical results with a specific seed. Tested on a Q4_0 7B LLaMA1 model.

No difference in speed that I can see, which isn't too surprising since like 90% of time is spent in matrix multiplication.