@li-plus 2023-08-27
> What is the motivation behind this change? In which situation does the dequantization performance actually make a difference?

@JohannesGaessler I'm optimizing [chatglm.cpp](https://github.com/li-plus/chatglm.cpp) and find that dequantization kernels cost ~50% of the total time of context computing. I only test short prompts though. Here is the nsys profile.

![image](https://github.com/ggerganov/llama.cpp/assets/39846316/4193df88-34aa-4382-b84e-a3e982031bfd)