@JohannesGaessler 2023-08-27
I would suggest you try a longer prompt, the `llama-bench` binary, or the `perplexity` binary. For long prompts I would be very surprised if it made a noticeable difference.