@JohannesGaessler 2023-08-27
When I tested it using the `llama-bench` binary 13.9% of the kernel time was spent in `dequantize_block` on master. But this is only with `-mmq 0` in conjunction with cuBLAS. The new mul_mat_q kernels which are both faster and use less VRAM do not invoke the corresponding dequantization kernel in the first place.