@jammm 2023-10-08
> I would suggest you try a longer prompt, the `llama-bench` binary, or the `perplexity` binary. For long prompts I would be very surprised if it made a noticeable difference.

noob question - what is the prompt length and content of the llama-bench program? Shouldn't the performance depend on what the prompt is asking?