@slaren 2023-10-08
`llama-bench` only measures the time to process a prompt, not the response. The length of the prompt can be configured with the `-p` parameter. The contents of the prompt are just `bos` tokens, but the performance should be the same regardless of the contents.