@jammm 2023-10-09
> `llama-bench` only measures the time to process a prompt, not the response. The length of the prompt can be configured with the `-p` parameter. The contents of the prompt are just `bos` tokens, but the performance should be the same regardless of the contents.

I see. Thanks! So the `tg` test simply calculates the amount of `bos` tokens based on the default number of tokens to be generated? Does this imply the same speed when running `main` with the same prompt length?