@jammm 2023-10-09
> The `tg` test simulates text generation by generating a sequence of the length given with `-n`, one token at a time. The token feed to the model is always `bos`, as in the `pp` tests. Essentially, it would be the same as if you ran `main` without a prompt and banned every token except `bos`, so that the only token that could be sampled is `bos`. The `pp` test should be equivalent to evaluating a prompt of the same length with `main`.

Thanks! That makes it a lot clearer now.