@ggerganov 2023-10-10
With the recent performance improvement of `-mmq 0` (#3412) this change looks more relevant.
Should we merge this and put some effort in optimizing the rest of the dequantization kernels?