@ggerganov 2023-10-10
OK. For future reference, adding a data point for low-batch decoding on V100 (#3479)

```
make -j && ./bin/llama-bench -m /mnt/llama.cpp/models/open-llama/7B-v2/ggml-model-f16.gguf -mmq 0 -n 0 -t 1 -p 1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,32,64,128,256,512
```

| model                   |   back       | ngl |         th |        mmq | test       |    master    t/s |    PR        t/s |    speedup |
| ----------------------- | ------------ | --: | ---------: | ---------: | ---------- | ---------------: | ---------------: | ---------: |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 1       |     45.40 ± 8.25 |     45.40 ± 8.11 |   1.000    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 2       |     24.78 ± 0.12 |     32.29 ± 0.18 |   1.303    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 3       |     36.65 ± 0.37 |     47.83 ± 0.25 |   1.305    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 4       |     49.36 ± 0.07 |     62.73 ± 1.64 |   1.271    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 5       |     60.40 ± 0.07 |     77.27 ± 0.83 |   1.279    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 6       |     72.84 ± 0.08 |     93.77 ± 0.18 |   1.287    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 7       |     82.90 ± 0.14 |    106.24 ± 0.11 |   1.282    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 8       |     96.52 ± 0.57 |    122.71 ± 3.46 |   1.271    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 9       |    106.96 ± 0.13 |    136.79 ± 0.09 |   1.279    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 10      |    119.88 ± 0.04 |    153.53 ± 0.65 |   1.281    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 11      |    129.90 ± 0.19 |    166.20 ± 0.25 |   1.279    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 12      |    143.29 ± 0.56 |    183.78 ± 0.24 |   1.283    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 13      |    151.01 ± 0.29 |    191.97 ± 0.17 |   1.271    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 14      |    165.57 ± 0.04 |    209.25 ± 3.43 |   1.264    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 15      |    175.41 ± 0.16 |    222.84 ± 0.66 |   1.270    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 16      |    190.85 ± 0.11 |    244.04 ± 0.24 |   1.279    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 32      |    337.25 ± 6.56 |   417.63 ± 10.76 |   1.238    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 64      |    575.15 ± 0.36 |    686.56 ± 0.43 |   1.194    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 128     |  1096.75 ± 38.20 |  1306.71 ± 31.41 |   1.191    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 256     |   1803.38 ± 1.90 |   2064.30 ± 6.43 |   1.145    |
| llama 7B Q4_0           |   CUDA       |  99 |          1 |          0 | pp 512     |   2329.00 ± 4.80 |   2538.12 ± 2.10 |   1.090    |