@slaren 2023-10-10
> Shouldn't that be handled by cublas/hipblas already? They should use tensor cores/WMMA.

Yes, but writing our own kernels would allow us to do this without having to dequantize the entire matrix to main memory first, and to use INT8 instead of FP16.